Bug fixes for Uploading files: Bug found that file choser window only load once Solved it by uing Event dispatching thread https://en.wikipedia.org/wiki/Event_dispatching_thread
