+++ src/main/java/seedu/address/logic/commands/UploadFileCommand.java
@@ -6,10 +6,11 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.files.FileStorage;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
-import seedu.address.storage.FileStorage;
+
 
 
 /**
@@ -41,7 +42,7 @@ public class UploadFileCommand extends Command {
 
         Person personToUpload = lastShownList.get(targetIndex.getZeroBased());
         FileStorage fileStorage = new FileStorage(personToUpload.getName().fullName);
-        new Thread(() -> fileStorage.uploadFile()).start();
+        fileStorage.uploadFile();
         return new CommandResult(String.format(MESSAGE_UPLOAD_SUCCESS, personToUpload));
     }
 }
