+++ src/main/java/seedu/address/model/ModelManager.java
@@ -11,8 +11,9 @@ import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
+import seedu.address.files.FileStorage;
 import seedu.address.model.person.Person;
-import seedu.address.storage.FileStorage;
+
 
 /**
  * Represents the in-memory model of the address book data.
@@ -99,8 +100,6 @@ public class ModelManager implements Model {
     @Override
     public void deletePerson(Person target) {
         addressBook.removePerson(target);
-        FileStorage fileStorage = new FileStorage(target.getName().fullName);
-        //new Thread(() -> imageStorage.uploadFile(target.getName().fullName)).start();
         FileStorage.deleteDrc(target.getName().fullName);
     }
 
@@ -114,7 +113,6 @@ public class ModelManager implements Model {
     @Override
     public void setPerson(Person target, Person editedPerson) {
         requireAllNonNull(target, editedPerson);
-
         addressBook.setPerson(target, editedPerson);
     }
